chore(package.json): update dependencies versions fix(footer.astro): update copyright information refactor(logos.astro): update heading text feat(navbar.astro): add theme switcher component add new file themeswitch.astro for the theme switcher component
The package.json file has been updated to include new versions of dependencies. This ensures that the project is using the latest and most stable versions of the packages.

In the footer.astro file, the copyright information has been updated to reflect the new company name, Akrista.

In the logos.astro file, the heading text has been changed to "Build with the best in tech" to better convey the message.

The navbar.astro file now includes a new component called Themeswitch, which allows users to switch between different themes (system, dark, and light). This provides a more customizable and user-friendly experience.

A new file, themeswitch.astro, has been added to the project. This file contains the implementation of the theme switcher component, which includes a dropdown menu with options for different themes and JavaScript logic to handle the theme switching functionality.